DSND-1599: Move FetchAppointmentsRequest to model dir
