+++ src/main/java/uk/gov/companieshouse/company_appointments/controller/CompanyAppointmentController.java
@@ -19,7 +19,7 @@ import uk.gov.companieshouse.company_appointments.exception.BadRequestException;
 import uk.gov.companieshouse.company_appointments.exception.NotFoundException;
 import uk.gov.companieshouse.company_appointments.exception.ServiceUnavailableException;
 import uk.gov.companieshouse.company_appointments.service.CompanyAppointmentService;
-import uk.gov.companieshouse.company_appointments.service.FetchAppointmentsRequest;
+import uk.gov.companieshouse.company_appointments.model.FetchAppointmentsRequest;
 import uk.gov.companieshouse.logging.Logger;
 import uk.gov.companieshouse.logging.LoggerFactory;
